Extracted base classes for ProfiledDbConnection and ProfiledDbCommand in the ServiceStack.Common assembly, so that they can be used outside of the ServiceStack environment.
