+++ src/ServiceStack/MiniProfiler/Data/ProfiledDbCommand.cs
@@ -3,19 +3,13 @@ using System.Data.Common;
 using System.Data;
 using System.Reflection;
 using System.Reflection.Emit;
-using ServiceStack.Data;
 
 #pragma warning disable 1591 // xml doc comments warnings
 
 namespace ServiceStack.MiniProfiler.Data
 {
-    public class ProfiledDbCommand : DbCommand, ICloneable, IHasDbCommand
+    public class ProfiledDbCommand : ProfiledCommand, ICloneable
     {
-        protected DbCommand _cmd;
-        protected DbConnection _conn;
-        protected DbTransaction _tran;
-        protected IDbProfiler _profiler;
-
         private bool bindByName;
         /// <summary>
         /// If the underlying command supports BindByName, this sets/clears the underlying
@@ -71,44 +65,13 @@ namespace ServiceStack.MiniProfiler.Data
 
 
         public ProfiledDbCommand(DbCommand cmd, DbConnection conn, IDbProfiler profiler)
+            : base(cmd, conn, profiler)
         {
-            if (cmd == null) throw new ArgumentNullException("cmd");
-
-            _cmd = cmd;
-            _conn = conn;
-
-            if (profiler != null)
-            {
-                _profiler = profiler;
-            }
-        }
-
-        public override string CommandText
-        {
-            get { return _cmd.CommandText; }
-            set { _cmd.CommandText = value; }
-        }
-
-        public override int CommandTimeout
-        {
-            get { return _cmd.CommandTimeout; }
-            set { _cmd.CommandTimeout = value; }
-        }
-
-        public override CommandType CommandType
-        {
-            get { return _cmd.CommandType; }
-            set { _cmd.CommandType = value; }
-        }
-
-        public IDbCommand DbCommand
-        {
-            get { return _cmd; }
         }
 
         protected override DbConnection DbConnection
         {
-            get { return _conn; }
+            get { return base.DbConnection; }
             set
             {
                 // TODO: we need a way to grab the IDbProfiler which may not be the same as the MiniProfiler, it could be wrapped
@@ -118,148 +81,42 @@ namespace ServiceStack.MiniProfiler.Data
                     _profiler = Profiler.Current;
                 }
 
-                _conn = value;
-                var awesomeConn = value as ProfiledDbConnection;
-                _cmd.Connection = awesomeConn == null ? value : awesomeConn.WrappedConnection;
+                base.DbConnection = value;
             }
         }
 
-        protected override DbParameterCollection DbParameterCollection
+        #region Wrapper properties for backwards compatibility
+
+        protected DbCommand _cmd
         {
-            get { return _cmd.Parameters; }
+            get { return DbCommand; }
+            set { DbCommand = value; }
         }
 
-        protected override DbTransaction DbTransaction
+        protected DbConnection _conn
         {
-            get { return _tran; }
-            set
-            {
-                this._tran = value;
-                var awesomeTran = value as ProfiledDbTransaction;
-                _cmd.Transaction = awesomeTran == null ? value : awesomeTran.WrappedTransaction;
-            }
+            get { return DbConnection; }
+            set { DbConnection = value; }
         }
 
-        public override bool DesignTimeVisible
+        protected DbTransaction _tran
         {
-            get { return _cmd.DesignTimeVisible; }
-            set { _cmd.DesignTimeVisible = value; }
+            get { return DbTransaction; }
+            set { DbTransaction = value; }
         }
 
-        public override UpdateRowSource UpdatedRowSource
+        protected IDbProfiler _profiler
         {
-            get { return _cmd.UpdatedRowSource; }
-            set { _cmd.UpdatedRowSource = value; }
+            get { return DbProfiler; }
+            set { DbProfiler = value; }
         }
 
-
-        protected override DbDataReader ExecuteDbDataReader(CommandBehavior behavior)
-        {
-            if (_profiler == null || !_profiler.IsActive)
-            {
-                return _cmd.ExecuteReader(behavior);
-            }
-
-            DbDataReader result = null;
-            _profiler.ExecuteStart(this, ExecuteType.Reader);
-            try
-            {
-                result = _cmd.ExecuteReader(behavior);
-                result = new ProfiledDbDataReader(result, _conn, _profiler);
-            }
-            catch (Exception e)
-            {
-                _profiler.OnError(this, ExecuteType.Reader, e);
-                throw;
-            }
-            finally
-            {
-                _profiler.ExecuteFinish(this, ExecuteType.Reader, result);
-            }
-            return result;
-        }
-
-        public override int ExecuteNonQuery()
-        {
-            if (_profiler == null || !_profiler.IsActive)
-            {
-                return _cmd.ExecuteNonQuery();
-            }
-
-            int result;
-
-            _profiler.ExecuteStart(this, ExecuteType.NonQuery);
-            try
-            {
-                result = _cmd.ExecuteNonQuery();
-            }
-            catch (Exception e)
-            {
-                _profiler.OnError(this, ExecuteType.NonQuery, e);
-                throw;
-            }
-            finally
-            {
-                _profiler.ExecuteFinish(this, ExecuteType.NonQuery, null);
-            }
-            return result;
-        }
-
-        public override object ExecuteScalar()
-        {
-            if (_profiler == null || !_profiler.IsActive)
-            {
-                return _cmd.ExecuteScalar();
-            }
-
-            object result;
-            _profiler.ExecuteStart(this, ExecuteType.Scalar);
-            try
-            {
-                result = _cmd.ExecuteScalar();
-            }
-            catch (Exception e)
-            {
-                _profiler.OnError(this, ExecuteType.Scalar, e);
-                throw;
-            }
-            finally
-            {
-                _profiler.ExecuteFinish(this, ExecuteType.Scalar, null);
-            }
-            return result;
-        }
-
-        public override void Cancel()
-        {
-            _cmd.Cancel();
-        }
-
-        public override void Prepare()
-        {
-            _cmd.Prepare();
-        }
-
-        protected override DbParameter CreateDbParameter()
-        {
-            return _cmd.CreateParameter();
-        }
-
-        protected override void Dispose(bool disposing)
-        {
-            if (disposing && _cmd != null)
-            {
-                _cmd.Dispose();
-            }
-            _cmd = null;
-            base.Dispose(disposing);
-        }
-
+        #endregion
 
         public ProfiledDbCommand Clone()
         { // EF expects ICloneable
             ICloneable tail = _cmd as ICloneable;
-            if (tail == null) throw new NotSupportedException("Underlying " + _cmd.GetType().GetOperationName() + " is not cloneable");
+            if (tail == null) throw new NotSupportedException("Underlying " + _cmd.GetType().FullName + " is not cloneable");
             return new ProfiledDbCommand((DbCommand)tail.Clone(), _conn, _profiler);
         }
         object ICloneable.Clone() { return Clone(); }
